requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
coverage==6.4.4
Flask==2.1.3
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-env==0.6.2
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
Werkzeug==2.1.2
